Overhauled entire code sense last commit...
A few functions weren't in a logical place - reorganized

Lots of temporary print statements, these will be removed in next version
- 99.999% sure they aren't necessary anymore as I finished debugging the issue that required them)
- This issue:  could not calibrate cage after starting the cage in a static test 

When calibrating, the plots used to only update with a new value when a new voltage was sent. This has been changed so now update_plot_time can be a smaller time increment than update_calibrate_time, which gives a much better visualization of what is happening. 

Next updates:
- get magnetometer data fed into the instruments class
- start looking into using a calibration file to allow a user to request a magnetic field 
- remove ALL print statements and switch it to the logging module (more professional, printing was always a temporary solution...)
- disable buttons based on what user does (ex: after a calibration is started, don't allow anything else to be pressed until finished)

Side note: Not many pushes recently due to a few major issues (and the fact that I've been the sole person interested in development). I'm very happy with where the code is now at. Assuming this method of calibrating works, this section of the GUI is complete.